+++ CHANGELOG.md
@@ -64,6 +64,7 @@ The format is based on [Keep a Changelog].
 
 ### Formatters
 
+* [dprint](https://dprint.dev) for various (depends on installed plugins).
 * [purs-tidy](https://github.com/natefaubion/purescript-tidy) for PureScript ([#182]).
 * [`jq`](https://stedolan.github.io/jq/) for
   [JSON](https://developer.mozilla.org/en-US/docs/Learn/JavaScript/Objects/JSON)